move `BarangayRecord`
